# backend/docs/API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: CozyVTT API
  description: |
    RESTful API for CozyVTT — a self-hosted, browser-based Virtual Tabletop platform.

    ## Authentication
    All endpoints (except `/health`, `/api/setup/*`, and public auth routes) require authentication via session cookies.
    Login via `/api/auth/login` to obtain a session cookie.

    ## MFA Flow
    If a user has MFA enabled, `/api/auth/login` returns `{ mfaRequired: true }` instead of a full session.
    The client must then call `/api/auth/mfa/verify-login` with a TOTP code or backup code to complete authentication.

    ## Authorization
    - **Platform roles**: `ADMIN` (full system access), `USER` (campaign access)
    - **Campaign roles**: `DM` (full campaign control), `PLAYER` (limited interaction), `SPECTATOR` (read-only)

    ## Rate Limiting
    - Global API limit: 300 requests per minute per IP (all `/api/*` routes)
    - Authentication endpoints: 5 requests per 15 minutes
    - MFA setup verification: 5 attempts per 15 minutes
    - MFA login verification: 5 attempts per 15 minutes (separate limit)
    - Asset uploads: 30 requests per minute per user (configurable via `ASSET_UPLOAD_RATE_LIMIT`)
    - Chat messages: 5 per minute (WebSocket)
    - Dice rolls: 30 per minute (WebSocket)
    - Token movement: 60 per second (WebSocket, throttled)

    Rate-limited responses return HTTP 429 with `RateLimit-*` standard headers.

    ## Real-time Events
...

  version: 1.0.0
  contact:
    name: CozyVTT
...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /api/assets/tokens/{id}:
    get:
      tags:
        - Assets
      summary: Serve token image
      operationId: serveToken
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Token image
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/campaigns/{campaignId}/maps/{id}/tokens:
    post:
      tags:
        - Maps
      summary: Add token to map
      description: Add a new token to the map (DM only)
      operationId: addToken
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenInput'
      responses:
        '201':
          description: Token added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    $ref: '#/components/schemas/Token'
                  map:
                    $ref: '#/components/schemas/Map'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/campaigns/{campaignId}/maps/{id}/tokens/{tokenId}:
    put:
      tags:
        - Maps
      summary: Update token
      description: Update token (DM can update any, players only tokens they control)
      operationId: updateToken
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                imageUrl:
                  type: string
                position:
                  type: object
                  properties:
                    x:
                      type: number
                    y:
                      type: number
                size:
                  type: object
                  properties:
                    width:
                      type: number
                    height:
                      type: number
                layer:
                  type: string
                  enum: [token, spirit]
                visible:
                  type: boolean
                controlledBy:
                  type: string
                  format: uuid
                  nullable: true
                rotation:
                  type: number
                conditions:
                  type: array
                  items:
                    type: string
                metadata:
                  type: object
      responses:
        '200':
          description: Token updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    $ref: '#/components/schemas/Token'
                  map:
                    $ref: '#/components/schemas/Map'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Maps
      summary: Delete token
      description: Remove token from map (DM only)
      operationId: deleteToken
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Token removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # WALLS, LIGHTS, & FOG
  # Per-map dynamic lighting and fog-of-war state.
  # ============================================
...
  /api/campaigns/{campaignId}/token-templates:
    get:
      tags: [Token Templates]
      summary: List token templates
      description: Get all token templates for a campaign. Supports search and type filtering.
      operationId: listTokenTemplates
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [npc, player, object]
      responses:
        '200':
          description: List of token templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/TokenTemplate'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Token Templates]
      summary: Create token template
      description: Create a new token template (DM only).
      operationId: createTokenTemplate
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenTemplateInput'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/campaigns/{campaignId}/token-templates/{id}:
    parameters:
      - $ref: '#/components/parameters/CampaignId'
      - $ref: '#/components/parameters/TemplateId'
    get:
      tags: [Token Templates]
      summary: Get token template
      operationId: getTokenTemplate
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Token template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTemplate'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Token Templates]
      summary: Update token template
      operationId: updateTokenTemplate
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenTemplateInput'
      responses:
        '200':
          description: Template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Token Templates]
      summary: Delete token template
      operationId: deleteTokenTemplate
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Template deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/campaigns/{campaignId}/token-templates/from-token:
    post:
      tags: [Token Templates]
      summary: Save token as template
      description: Create a token template from an existing map token's properties.
      operationId: saveTokenAsTemplate
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenTemplateInput'
      responses:
        '201':
          description: Template created from token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/campaigns/{campaignId}/token-templates/{id}/copy-to/{targetCampaignId}:
    post:
      tags: [Token Templates]
      summary: Copy template to another campaign
      description: |
        Copy a token template to another campaign. The authenticated user must
        be the DM of **both** the source and target campaigns.
      operationId: copyTokenTemplate
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - $ref: '#/components/parameters/TemplateId'
        - name: targetCampaignId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Template copied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenTemplate'
        '400':
          description: Cannot copy a template to the same campaign.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # CAMPAIGN EXPORT/IMPORT ENDPOINTS
  # ============================================
...
  /api/admin/smtp/test:
    post:
      tags: [Admin]
      summary: Send a test email to verify SMTP configuration
      operationId: adminTestSmtp
      description: Sends a test email to the calling admin's own email address. Returns 400 if SMTP is not configured via environment variables.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Test email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: SMTP is not configured (missing env vars)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          description: Failed to send email (SMTP connection error)
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cozyvtt.sid
      description: Session cookie obtained via /api/auth/login
  parameters:
    CampaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Campaign UUID
...
    TemplateId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Token template UUID
...
  schemas:
...
    Map:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campaignId:
          type: string
          format: uuid
        name:
          type: string
        imageUrl:
          type: string
        baseLayerUrl:
          type: string
        spiritLayerUrl:
          type: string
          nullable: true
          description: Hidden from non-DM users when spirit layer is disabled
        width:
          type: integer
        height:
          type: integer
        gridSize:
          type: integer
          default: 50
        tokens:
          type: array
          items:
            $ref: '#/components/schemas/Token'
          description: Filtered by role (spirit layer tokens hidden from players)
        annotations:
          type: array
          items:
            type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TokenHp:
      type: object
      properties:
        current:
          type: integer
        max:
          type: integer
        temp:
          type: integer
    Token:
      type: object
      properties:
        id:
          type: string
          format: uuid
        characterId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        imageUrl:
          type: string
        position:
          type: object
          required:
            - x
            - y
          properties:
            x:
              type: number
            y:
              type: number
        size:
          type: object
          required:
            - width
            - height
          properties:
            width:
              type: number
            height:
              type: number
        layer:
          type: string
          enum: [token, spirit]
          description: 'token = normal layer; spirit = spirit/astral layer (hidden from players by default)'
        visible:
          type: boolean
          description: If false, hidden from players (DM can still see)
        controlledBy:
          type: string
          format: uuid
          nullable: true
          description: User ID of the player who controls this token (null = DM-only)
        rotation:
          type: number
          default: 0
          description: Rotation in degrees
        conditions:
          type: array
          items:
            type: string
          description: Active status conditions (e.g. "Poisoned", "Prone")
        metadata:
          type: object
          description: Arbitrary custom data
        type:
          type: string
          enum: [player, npc, object]
          description: Token classification; affects context menu options and initiative behavior
        disposition:
          type: string
          enum: [friendly, neutral, hostile]
          nullable: true
          description: NPC attitude toward the party
        hp:
          $ref: '#/components/schemas/TokenHp'
          nullable: true
          description: Hit points / health for this token
        showHpBar:
          type: boolean
          default: false
          description: Whether to display the HP bar on the map canvas
        notes:
          type: string
          description: DM-only notes for this token
        initiative:
          type: integer
          nullable: true
          description: Initiative roll result for combat turn ordering
    TokenInput:
      type: object
      required:
        - name
        - imageUrl
        - position
      properties:
        characterId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        imageUrl:
          type: string
        position:
          type: object
          required:
            - x
            - y
          properties:
            x:
              type: number
            y:
              type: number
        size:
          type: object
          properties:
            width:
              type: number
              default: 1
            height:
              type: number
              default: 1
        layer:
          type: string
          enum: [token, spirit]
          default: token
        visible:
          type: boolean
          default: true
        controlledBy:
          type: string
          format: uuid
          nullable: true
        rotation:
          type: number
          default: 0
        conditions:
          type: array
          items:
            type: string
        metadata:
          type: object
        type:
          type: string
          enum: [player, npc, object]
          default: npc
        disposition:
          type: string
          enum: [friendly, neutral, hostile]
          nullable: true
        hp:
          $ref: '#/components/schemas/TokenHp'
          nullable: true
        showHpBar:
          type: boolean
          default: false
        notes:
          type: string
          default: ''
        initiative:
          type: integer
          nullable: true
...
    TokenTemplate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        type:
          type: string
          enum: [npc, player, object]
        disposition:
          type: string
          nullable: true
        displayMode:
          type: string
          enum: [pog, top-down, full-art]
        size:
          type: object
          properties:
            width:
              type: integer
            height:
              type: integer
        notes:
          type: string
          nullable: true
        hp:
          type: object
          nullable: true
          properties:
            current:
              type: integer
            max:
              type: integer
            temp:
              type: integer
        showHpBar:
          type: boolean
        statBlock:
          type: object
          nullable: true
        sightRadius:
          type: number
          nullable: true
        campaignId:
          type: string
        createdById:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TokenTemplateInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        imageUrl:
          type: string
        type:
          type: string
          enum: [npc, player, object]
        disposition:
          type: string
        displayMode:
          type: string
          enum: [pog, top-down, full-art]
        size:
          type: object
          properties:
            width:
              type: integer
            height:
              type: integer
        notes:
          type: string
        hp:
          type: object
          nullable: true
          properties:
            current:
              type: integer
            max:
              type: integer
            temp:
              type: integer
        showHpBar:
          type: boolean
        statBlock:
          type: object
          nullable: true
        sightRadius:
          type: number
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: BadRequest
          description: Short machine-readable error identifier
        message:
          type: string
          example: One or more fields failed validation.
          description: Human-readable error description
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: Optional structured details (e.g. per-field Zod issues)
...
  responses:
    BadRequest:
      description: Bad Request - Invalid input or validation failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BadRequest
            message: Required field "name" is missing.
    Unauthorized:
      description: Unauthorized - Not authenticated or MFA pending.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Authentication required.
    Forbidden:
      description: Forbidden - Authenticated but insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: You do not have permission to perform this action.
    NotFound:
      description: Not Found - Resource does not exist or you cannot see it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NotFound
            message: Campaign not found.
...